backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"alembic==1.12.0",
	"beautifulsoup4==4.12.2",
	"blinker==1.6.2",
	"certifi==2023.7.22",
	"cffi==1.16.0",
	"charset-normalizer==3.3.0",
	"click==8.1.7",
	"cryptography==43.0.0",
	"flask==2.3.3",
	"flask-cors==4.0.0",
	"flask-mail==0.10.0",
	"flask-migrate==4.0.5",
	"flask-qrcode>=3.2.0",
	"flask-sqlalchemy==3.0.3",
	"greenlet==3.0.0",
	"html5lib==1.1",
	"idna==3.4",
	"importlib-metadata==6.8.0",
	"itsdangerous==2.1.2",
	"jinja2==3.1.2",
	"mako==1.2.4",
	"markupsafe==2.1.3",
	"pdfkit>=1.0.0",
	"psycopg2-binary==2.9.6",
	"pycparser==2.22",
	"pyjwt==2.8.0",
	"python-dateutil==2.9.0.post0",
	"python-decouple==3.8",
	"python-dotenv==1.0.0",
	"requests==2.31.0",
	"six==1.16.0",
	"soupsieve==2.5",
	"sqlalchemy==2.0.21",
	"typing-extensions==4.8.0",
	"urllib3==2.0.6",
	"webencodings==0.5.1",
	"werkzeug==3.0.0",
	"zipp==3.17.0",
]
